Fix qr capture pref: activity and fragment issue
