+++ app/src/main/java/com/cmput301w23t47/canary/view/fragment/QrCapturePreferenceFragment.java
@@ -1,54 +1,50 @@
-package com.cmput301w23t47.canary;
-
+package com.cmput301w23t47.canary.view.fragment;
 
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.CompoundButton;
-import android.widget.EditText;
 
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.cmput301w23t47.canary.R;
 
 public class QrCapturePreferenceFragment extends Fragment {
 
-    static void takeSnapshot() {
+    public QrCapturePreferenceFragment() {}
 
-    }
-
-    static void noSnapshot() {
-
-    }
-
-    public QrCapturePreferenceFragment() {
-        super(R.layout.qr_details);
-    }
-
-    private Button takeSnap;
-    private Button noSnap;
-    private CompoundButton checkBox;
+//    /**
+//     * Use this factory method to create a new instance of
+//     * this fragment using the provided parameters.
+//     *
+//     * @param param1 Parameter 1.
+//     * @param param2 Parameter 2.
+//     * @return A new instance of fragment BlankFragment.
+//     */
+//    // TODO: Rename and change types and number of parameters
+//    public static QrCapturePreferenceFragment newInstance(String param1, String param2) {
+//        QrCapturePreferenceFragment fragment = new QrCapturePreferenceFragment();
+//        Bundle args = new Bundle();
+//        args.putString(ARG_PARAM1, param1);
+//        args.putString(ARG_PARAM2, param2);
+//        fragment.setArguments(args);
+//        return fragment;
+//    }
 
     @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        View view = LayoutInflater.from(getContext()).inflate(R.layout.qr_details, null);
-        EditText qrName = view.findViewById(R.id.qr_name);
-        EditText qrScore = view.findViewById(R.id.qr_score);
-
-        takeSnap.setOnClickListener(v -> {
-            takeSnapshot();
-        });
-        noSnap.setOnClickListener(v -> {
-            noSnapshot();
-        });
-
-
+        if (getArguments() != null) {
+//            mParam1 = getArguments().getString(ARG_PARAM1);
+//            mParam2 = getArguments().getString(ARG_PARAM2);
+        }
     }
-    public void onCheckBoxClicked(View view) {
-        boolean checked = ((CheckBox) view).isChecked();
 
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_qr_capture_preference, container, false);
     }
 }
