+++ app/src/main/res/layout/fragment_home.xml
@@ -26,13 +26,4 @@ tools:context=".view.fragment.HomeFragment"
         app:layout_constraintTop_toBottomOf="@+id/textView"
         android:text="@string/scanQr"
         />
-
-<!--    <fragment-->
-<!--        android:name="androidx.navigation.fragment.NavHostFragment"-->
-<!--        android:id="@+id/fragment_container_view_home"-->
-<!--        android:layout_width="match_parent"-->
-<!--        android:layout_height="match_parent"-->
-<!--        app:defaultNavHost="true"-->
-<!--        app:navGraph="@navigation/home_nav_graph"-->
-<!--        />-->
 </androidx.constraintlayout.widget.ConstraintLayout>
