+++ app/src/main/res/layout/fragment_qr_capture_preference.xml
@@ -1,13 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    android:background="#DCD9D9"
-    tools:context=".MainActivity">
-
+    tools:context=".view.fragment.QrCapturePreferenceFragment">
     <LinearLayout
         android:layout_width="300dp"
         android:layout_gravity="center"
@@ -116,6 +113,5 @@
         tools:layout_editor_absoluteX="68dp"
         tools:layout_editor_absoluteY="650dp" />
 
+
 </LinearLayout>
-
-
