+++ app/src/main/res/navigation/add_qr_nav_graph.xml
@@ -3,13 +3,20 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
+    app:startDestination="@id/addNewQrFragment">
 
+    <activity
+        android:id="@+id/scanQRCodeActivity"
+        android:name="com.cmput301w23t47.canary.view.activity.ScanQRCodeActivity"
+        android:label="ScanQRCodeActivity" />
+    <activity
+        android:id="@+id/captureSnapshotActivity"
+        android:name="com.cmput301w23t47.canary.view.activity.CaptureSnapshotActivity"
+        android:label="CaptureSnapshotActivity" />
     <fragment
-        android:id="@+id/FirstFragment"
+        android:id="@+id/addNewQrFragment"
         android:name="com.cmput301w23t47.canary.view.fragment.AddNewQrFragment"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_add_new_qr">
+        android:label="AddNewQrFragment" >
         <action
             android:id="@+id/action_scanQrCode"
             app:destination="@id/scanQRCodeActivity" />
@@ -17,20 +24,9 @@
             android:id="@+id/action_qrCapturePreference"
             app:destination="@id/qrCapturePreferenceFragment" />
     </fragment>
-    <activity
-        android:id="@+id/scanQRCodeActivity"
-        android:name="com.cmput301w23t47.canary.view.activity.ScanQRCodeActivity"
-        android:label="ScanQRCodeActivity" />
     <fragment
         android:id="@+id/qrCapturePreferenceFragment"
-        android:name="com.cmput301w23t47.canary.QrCapturePreferenceFragment"
-        android:label="QrCapturePreferenceFragment" >
-        <action
-            android:id="@+id/action_captureSnapshot"
-            app:destination="@id/captureSnapshotActivity" />
-    </fragment>
-    <activity
-        android:id="@+id/captureSnapshotActivity"
-        android:name="com.cmput301w23t47.canary.view.activity.CaptureSnapshotActivity"
-        android:label="CaptureSnapshotActivity" />
+        android:name="com.cmput301w23t47.canary.view.fragment.QrCapturePreferenceFragment"
+        android:label="fragment_qr_capture_preference"
+        tools:layout="@layout/fragment_qr_capture_preference" />
 </navigation>
